requirements.txt
django==2.2.16
djangorestframework==3.12.4
pytest-django==3.8.0
pytest-pythonpath==0.7.3
pytest==5.3.5
six==1.14.0